Boss fight mostly done
World 2 level picker under way as well!

+++ scripts/knockback_floppy/knockback_floppy.gml
@@ -1,8 +1,12 @@
 // Script assets have changed for v2.3.0 see
-function knockback_floppy(){
-		global.shake = true;
-		oHero.knockingback = true;
-		oHero.ySpeed -= 4;
-		oHero.alarm[2] = 10; //alarm to kill floppy in 3 seconds
 
+	function knockback_floppy(){
+		if (oHero.knockingback == false){		
+		
+			global.shake = true;
+			oHero.knockingback = true;
+			oHero.ySpeed -= 4;
+			oHero.alarm[2] = 10; //alarm to kill floppy in 3 seconds
+
+	}
 }
